# Conflicts:
#	.circleci/config.yml
#	src/sdp_solve/SDP_Solver/run/run.cxx
#	src/sdpb/solve.cxx
#	test/src/integration_tests/util/diff_sdpb_out.cxx

+++ src/sdp_solve/Solver_Parameters/String_To_Bytes_Translator.hxx
@@ -24,7 +24,7 @@ struct String_To_Bytes_Translator
   static size_t from_string(const std::string &s)
   {
     std::istringstream iss(s);
-    size_t number;
+    double number;
     std::string suffix;
 
     iss >> number >> suffix;
